pyproject.toml
[tool.poetry]
name = "cprag"
version = "0.1.0"
description = ""
authors = ["max"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
uvicorn = "^0.32.0"
jinja2 = "^3.1.4"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
greenlet = "^3.1.1"
python-docx = "^1.1.2"
transformers = "^4.46.2"
python-magic = "^0.4.27"
python-multipart = "^0.0.17"
pdf2image = "^1.17.0"
bm25s = "^0.2.3"
easyocr = "^1.7.2"
vllm = "^0.6.3.post1"
openai = "^1.54.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
